navigation 추상화 진행 1. navigation과 관련된 로직들 hook으로 분리하여 관리
